requirements.txt
black
celluloid
gensim
graphviz
grip
isort
ipywidgets
jupyterlab
jupyterlab-code-formatter
loguru
line-profiler
matplotlib
numpy
pandas
plotly
pyarrow
scipy
seaborn
sklearn
statsmodels
tensorboard
tensorflow
torch
tqdm
PyYAML
